float

+++ src/VecSim/spaces/IP/IP_AVX512F_SQ8.h
@@ -40,7 +40,7 @@ template <unsigned char residual> // 0..15
 float SQ8_InnerProductImp(const void *pVec1v, const void *pVec2v, size_t dimension, float inv_norm = 1.0f) {
     const float *pVec1 = static_cast<const float *>(pVec1v);
     const uint8_t *pVec2 = static_cast<const uint8_t *>(pVec2v);
-    const uint8_t *pEnd1 = pVec1 + dimension;
+    const float *pEnd1 = pVec1 + dimension;
 
     // Get dequantization parameters from the end of pVec2
     const float min_val = *reinterpret_cast<const float *>(pVec2 + dimension);
